Split 2002x-operand-formats test
My original goal was to add a sign-extended decimal format, but that
turned out to be awkward.  It works for data items and instructions
with immediate operands (e.g. "LDA #-1"), but is either wrong or
useless for address operands, since most assemblers treat integers
as 32-bit values.  (LDA -1 is not LDA $FFFF, it's LDA $FFFFFFFF,
which is not useful unless your asm is doing an implicit mod.)

There's also a bit of variability in how assemblers treat negative
values, so I'm shelving the idea for now.  I'm keeping the updated
tests, which are now split into 6502 / 65816 parts.

Also, updated the formatter to output all decimal values as unsigned.
Most assemblers were fine with negative values, but 64tass .dword
insists on positive.  Rather than make the opcode conditional on the
value's range, we now just always output unsigned decimal, which
all current assemblers accept.

+++ Asm65/Formatter.cs
@@ -535,12 +535,12 @@ namespace Asm65 {
         }
 
         /// <summary>
-        /// Formats a value as decimal.
+        /// Formats a 32-bit integer value as unsigned decimal.
         /// </summary>
         /// <param name="value">Value to convert.</param>
         /// <returns>Formatted string.</returns>
         public string FormatDecimalValue(int value) {
-            return value.ToString();
+            return ((uint)value).ToString();
         }
 
         /// <summary>
